+++ packages/iov-core/src/multichainsigner.spec.ts
@@ -53,6 +53,7 @@ describe("MultiChainSigner", () => {
     const signer = new MultiChainSigner(profile);
     expect(signer).toBeTruthy();
     expect(signer.chainIds().length).toEqual(0);
+    signer.shutdown();
   });
 
   // This uses setup from iov-bns...
@@ -118,6 +119,8 @@ describe("MultiChainSigner", () => {
         throw new Error("Unexpected transaction kind");
       }
       expect(last.memo).toEqual(memo);
+
+      signer.shutdown();
     });
 
     it("can add two chains", async () => {
@@ -166,6 +169,8 @@ describe("MultiChainSigner", () => {
       const account2 = await connection2.getAccount({ address: ganacheAddr });
       expect(account2).toBeDefined();
       expect(account2!.balance.length).toEqual(1);
+
+      signer.shutdown();
     });
   });
 
@@ -195,5 +200,45 @@ describe("MultiChainSigner", () => {
       .addChain(invalidConnector)
       .then(() => fail("must not resolve"))
       .catch(error => expect(error).toMatch(/connected chain ID does not match/i));
+
+    signer.shutdown();
+    signer2.shutdown();
+    signer3.shutdown();
+  });
+
+  describe("isValidAddress", () => {
+    it("can use isValidAddress for BNS and Ethereum", async () => {
+      pendingWithoutBnsd();
+      pendingWithoutEthereum();
+
+      const signer = new MultiChainSigner(new UserProfile());
+
+      const bnsConnection = (await signer.addChain(bnsConnector(bnsdTendermintUrl))).connection;
+      const ethereumConnection = (await signer.addChain(ethereumConnector(httpEthereumUrl, undefined)))
+        .connection;
+
+      const bnsChainId = bnsConnection.chainId();
+      const ethereumChainId = ethereumConnection.chainId();
+
+      // valid
+      expect(signer.isValidAddress(bnsChainId, "tiov142424242424242424242424242424242vmucnv")).toEqual(true);
+      expect(signer.isValidAddress(ethereumChainId, "0x890b61ca61fa5b5336bb3ec142fa0da250592337")).toEqual(
+        true,
+      );
+
+      // invalid
+      expect(signer.isValidAddress(bnsChainId, "")).toEqual(false);
+      expect(signer.isValidAddress(ethereumChainId, "")).toEqual(false);
+      expect(signer.isValidAddress(bnsChainId, "123")).toEqual(false);
+      expect(signer.isValidAddress(ethereumChainId, "123")).toEqual(false);
+
+      // wrong chains
+      expect(signer.isValidAddress(ethereumChainId, "tiov142424242424242424242424242424242vmucnv")).toEqual(
+        false,
+      );
+      expect(signer.isValidAddress(bnsChainId, "0x890b61ca61fa5b5336bb3ec142fa0da250592337")).toEqual(false);
+
+      signer.shutdown();
+    });
   });
 });